backend/requirements.txt
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2

# Core
Django>=5.0,<6.0
djangorestframework>=3.15

# Autenticación y seguridad
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# Google OAuth
google-auth
requests

# Entorno
python-dotenv

# Imágenes
Pillow

# Interfaz de administración moderna
django-jazzmin



dj-database-url
python-dotenv
psycopg2-binary